Remove unused device fields (#25)
This commit removes fields that are not directly used by the terraform provider from
the `Device` type.

This change is required to fix #24, as the API returns blank strings for dates, resulting in
errors unmarshalling the JSON response into a `Device` type. Since it's unlikely the provider
will ever need these dates, we can just remove them. The provider also only exposes the id
and name, so everything else can be removed for now too.

Signed-off-by: David Bond <davidsbond93@gmail.com>