Support hujson for ACL files
This commit swaps all uses of `encoding/json` with `tailscale/hujson`. This allows users of the provider
to have a hujson file locally that contains all the features such as comments and trailing commas.

When the ACL is pushed to tailscale via the provider, the comments will be lost when encoding/decoding
is done into the `ACL` type. However, the source of truth for the ACL will be local along with the terraform
configuration files, so this should not be a problem.

Closes #17

Signed-off-by: David Bond <davidsbond93@gmail.com>