switch resource type for DeviceSubnetRoutes.routes to Set
TypeList requires the items to be in a specific order; TypeSet does not. Using TypeList can result in unnecessary change detection if output returned by TS API (which does not appear to return in a fully deterministic or sorted order) is in a different order than input. It also makes diffs more difficult to read if e.g. you add a route mid-list and it detects all routes after that as changed.

In my testing, at least, this change appears to migrate cleanly from the previous version with TypeList.