Handle disappearing one-off authentication keys
This commit modifies the `tailnet_key` resource slightly so that one-off keys no longer being present
will not result in errors.

To do this, deletions that result in a 404 error will no longer return an error. Any other errors returned
via the API will still impact it. When reading keys, if a 404 error is returned and the key is not reusable
within the state, it will no longer return an error.

Closes #59

Signed-off-by: David Bond <davidsbond93@gmail.com>