Add wait_for argument to tailscale_device data source (#72)
This commit adds a new argument to the `tailscale_device` data source that allows for the configuration
of a retry mechanism. If this argument is provided, the provider will retry obtaining device data every
second up until the maximum duration specified in the `wait_for` argument.

This argument is optional, if it is not specified then the first error will fail the read. The field
is expected to use the string representation of go's `time.Duration` type.

Closes #69 (nice)

Signed-off-by: David Bond <davidsbond93@gmail.com>