Add tailscale_device_key resource
Closes #78

This commit introduces the `tailscale_device_key` resource which is used to modify the properties of the device's key.

Currently, this resource can disable device key expiry and pre-authorize a device for the tailnet. The read operation
checks the value of the device's `authorized` and `keyExpiryDisabled` fields, as there's no way to read the key properties
directly.

Signed-off-by: David Bond <davidsbond93@gmail.com>