Error when using an unsupported ACL field
Closes #122
Related to #121

This commit modifies the ACL decoding function to use `json.Decoder` and enables
the `DisallowUnknownFields` flag. This means that when someone using this provider
is trying to use an ACL feature that is not yet supported by the provider, they
will get an error prompting them to raise an issue on GitHub.

Initially, these errors would fail silently and users would be surprised to find
whatever unsupported field they used did not make its way to the Talscale API.

Signed-off-by: David Bond <davidsbond93@gmail.com>