Add support for terraform import
This commit is an initial attempt to set up support for the `terraform import`
command. It includes the default fallthrough importer for global resources
such as the ACL, DNS preferences etc.

For device-specific resources, the name of the device can be provided which
will be resolved to a device identifier. The device identifier is then
added to the resource so that a refresh will include it.

Signed-off-by: David Bond <davidsbond93@gmail.com>