tailscale: fix default User-Agent header to include terraform version
The `schema.Provider.UserAgent` helper function was previously being
called at a point in the plugin lifecycle before the `TerraformVersion`
for the provider was set by Terraform Core. This was resulting in the
Terraform version information being missing from the User-Agent header.
The call to `schema.Provider.UserAgent` has now been moved into the
`providerConfigure` func as this is called later in the configure
lifecycle where the Terraform version is properly set.

Fixes https://github.com/tailscale/corp/issues/19630

Signed-off-by: Mario Minardi <mario@tailscale.com>