tailscale: pass base URL to OAuthConfig
Pass the `parsedBaseURL` to `tsclient.OAuthConfig`. The base URL the
OAuth client used for attempting to generate tokens was always the
default URL previously which causes issues if the base URL for the
provider and the rest of the tailscale client is set to something else.

Fixes https://github.com/tailscale/terraform-provider-tailscale/issues/438

Signed-off-by: Mario Minardi <mario@tailscale.com>