tailscale: add support for S3 logstreaming
We recently added support for S3 logstreaming endpoints to our API. This
involved adding several new fields to the LogstreamConfiguration resource
(available when the new "s3" destinationType is used), plus a new
AWSExternalID resource needed when AWS role-based authentication is used
for an S3 logstreaming endpoint.

This commit updates the Terraform provider to reflect these changes. We add
support for the new fields to the tailscale_logstream_configuration
resource, and create a new tailscale_aws_external_id resource.

Unfortunately, we have to add some special handling for the "user" field of
tailscale_logstream_configuration. Previously, we specified a default dummy
value of "user" for the "user" field. Now, the PUT LogstreamConfiguration
endpoint does not allow any value to be specified for a s3 destinationType.
We could loosen that restriction at the API layer, but I would prefer not
to do that. However, we don't want to break any existing users who are
relying on this default and have it in their Terraform state. So the
approach I've chosen is to add special handling for the "user" field:
- When creating or updating a configuration, if destinationType == "s3" and
  user == "user", we assume the "user" value came from the default and we
  do not actually send it to the API.
- When reading or importing a configuration, if destinationType == "s3" and
  there is no user field (which should always be true), we set
  user = "user" in the Terraform state.

Fixes #458
Fixes tailscale/corp#24533

Signed-off-by: Zach Hauser <zehauser@gmail.com>