tailscale: add node_id support
- Add `node_id` to both `tailscale_device` and `tailscale_devices` data
sources. This is already exposed by the Go client and something we want
to provide to users.
- Update docs to indicate that `id` is a legacy identifier and node_id
should be preferred.

- In `resource_device_authorization.go`, detect if the user passes a
node_id into device_id and save it to state instead of the legacy id,
this prevents a perpetual diff.
- Add backwards-compatibility and node_id tests in
`resource_device_authorization_test.go`

- In `resource_device_key.go`, detect if the user passes a node_id into
device_id and save it to state instead of the legacy id,
this prevents a perpetual diff.
- Add backwards compatibility and node_id tests in
`resource_device_key_test.go`

- In `resource_device_tags.go`, detect if the user passes a node_id into
device_id and save it to state instead of the legacy id,
this prevents a perpetual diff.
- Add backwards compatibility and node_id tests in
`resource_device_tags_test.go`

- Add a test in `resource_device_subnet_routes_test.go` to verify
backwards compatibility that `node_id` works as `device_id`. Since we
use a UUID for this resource, we don't need to update the read step as
with the others.

Long term, we will likely want to switch users over to using `node_id`
as the resource id, but will want to do that during a major version
change.

Signed-off-by: Zach Buchheit <zachb@tailscale.com>