all: update docs for ACL resource and data source
Update the documentation for the ACL resource and data source to clarify
that they control / read the entire policy file for a tailnet and not
just the ACLs section of the policy file.

Update the example usage for the resource to show usage of grants over
the deprecated acl syntax.

Updates https://github.com/tailscale/terraform-provider-tailscale/issues/514

Signed-off-by: Mario Minardi <mario@tailscale.com>